requirements.txt
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
google-api-core==1.30.0
google-auth==1.32.0
google-cloud-vision==2.3.2
googleapis-common-protos==1.53.0
grpcio==1.38.1
idna==2.10
numpy==1.21.0
packaging==20.9
pandas==1.2.5
proto-plus==1.18.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
urllib3==1.26.5
